+++ srcs/buffer_util.c
@@ -6,7 +6,7 @@
 /*   By: plam <plam@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/26 12:46:24 by plam              #+#    #+#             */
-/*   Updated: 2020/02/07 13:30:45 by plam             ###   ########.fr       */
+/*   Updated: 2020/02/08 12:18:08 by plam             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	ft_bzero(void *s, size_t n)
 		str[i++] = 0;
 }
 
-void	buffer_register(t_print *printer) // problème avec le nombre des espaces
+void	buffer_register(t_print *printer)
 {
 	size_t	i;
 	size_t	sp;
@@ -32,7 +32,6 @@ void	buffer_register(t_print *printer) // problème avec le nombre des espaces
 	i = 0;
 	z = set_zeros(printer);
 	sp = set_spaces(printer);
-	//printf("buff = %s, z = %zu, sp = %zu\n", printer->buff, z, sp);
 	while (printer->buff[i])
 		i++;
 	while (sp-- > 0)
@@ -46,13 +45,13 @@ void	buffer_register(t_print *printer) // problème avec le nombre des espaces
 	ft_putstr(printer->buff);
 }
 
-/*faire une fonction aui ajoute des espacements comme précédemment mais pour les string*/
-
 size_t	len_add(t_print *printer)
 {
-	if (printer->acc >= printer->size && printer->acc >= printer->index)
+	if (printer->acc >= (int)printer->size
+		&& printer->acc >= (int)printer->index)
 		return (printer->acc);
-	else if (printer->size >= printer->acc && printer->size >= printer->index)
+	else if ((int)printer->size >= printer->acc &&
+			printer->size >= printer->index)
 		return (printer->size);
 	else
 		return (printer->index);
